Moved NameAndAccountType to Presentation layer.
