+++ app/src/main/java/com/sycosoft/allsee/presentation/mappers/NameAndAccountTypeMapper.kt
@@ -1,6 +1,6 @@
 package com.sycosoft.allsee.presentation.mappers
 
-import com.sycosoft.allsee.domain.models.NameAndAccountType
+import com.sycosoft.allsee.presentation.models.NameAndAccountType
 import com.sycosoft.allsee.domain.models.Person
 
 object NameAndAccountTypeMapper {
