+++ app/src/main/java/com/sycosoft/allsee/presentation/models/NameAndAccountType.kt
@@ -1,4 +1,4 @@
-package com.sycosoft.allsee.domain.models
+package com.sycosoft.allsee.presentation.models
 
 data class NameAndAccountType(
     val name: String,
