+++ app/src/test/java/com/sycosoft/allsee/presentation/usecases/GetNameAndAccountTypeUseCaseTests.kt
@@ -1,6 +1,6 @@
 package com.sycosoft.allsee.presentation.usecases
 
-import com.sycosoft.allsee.domain.models.NameAndAccountType
+import com.sycosoft.allsee.presentation.models.NameAndAccountType
 import com.sycosoft.allsee.domain.models.Person
 import com.sycosoft.allsee.domain.models.types.AccountHolderType
 import junit.framework.TestCase.assertEquals
